+++ config/.env.sample
@@ -17,28 +17,39 @@ LOG_LEVEL='INFO'
 # The .srt files must be stored within this folder.
 VIDEO_TO_USE = ''
 
+# Specify the type of question generation model to use.
+# Available options: 'BERTopic', 'LangChain'
+# Defaults to 'BERTopic'.
+GENERATION_MODEL = 'BERTopic'
+
 # Specify the number of questions that will be attempted to be generated for the given video.
-# There might not be enough clear context to generate the specified number of questions, 
-# So the actual number of questions generated might be less, equal to number of relevant regions found useful to generate questions.
-# Defaults to -1, which will generate one question per topic.
-QUESTION_COUNT = -1
+# In BERTopic mode, there might not be enough clear context to generate the specified number of questions, 
+#   So the actual number of questions generated might be less in certain cases,
+#   which will equal the number of relevant regions found useful to generate questions.
+#   If set to -1, the number of questions generated will be equal to the number of topics found.
+# In LangChain mode, the number of questions generated will be equal to the specified number.
+# Defaults to 3 questions generated per video.
+QUESTION_COUNT = 3
 
 # Toggle for overwriting existing saved transcript if present, to reprocess and extract the transcript.
 # 1 for overwriting existing models, 0 for not overwriting existing models.
 # Defaults to 0.
 OVERWRITE_EXISTING_TRANSCRIPT = 0
 
-# Toggle for overwriting existing saved BERTopic models if present, to run a new model.
-# NOTE: If you choose to overwrite the existing transcript data, this will be automatically set to 1 as well.
-# 1 for overwriting existing models, 0 for not overwriting existing models.
-# Defaults to 0.
-OVERWRITE_EXISTING_TOPICMODEL = 0
-
 # Toggle for overwriting existing saved question data if present, to generate new questions.
 # NOTE: If you choose to overwrite the existing topic model, this will be automatically set to 1 as well.
 # 1 for overwriting existing question data, 0 for not overwriting existing question data.
 # Defaults to 0
 OVERWRITE_EXISTING_QUESTIONS = 0
+
+#-------------------------Parameters For BERTopic only-------------------------#
+# The following parameters are only used when the 'BERTopic' mode is selected for question generation.
+
+# Toggle for overwriting existing saved BERTopic models if present, to run a new model.
+# NOTE: If you choose to overwrite the existing transcript data, this will be automatically set to 1 as well.
+# 1 for overwriting existing models, 0 for not overwriting existing models.
+# Defaults to 0.
+OVERWRITE_EXISTING_TOPICMODEL = 0
 
 # Approximate window size of the captions taken from the .srt file.
 # Value must be >= 1, but it is recommended to keep it between 10-30.
